+++ src/createRoom.js
@@ -37,17 +37,11 @@ function createRoom(opts) {
     opts = opts || {};
 
     const ErrorDialog = sdk.getComponent("dialogs.ErrorDialog");
-    const NeedToRegisterDialog = sdk.getComponent("dialogs.NeedToRegisterDialog");
     const Loader = sdk.getComponent("elements.Spinner");
 
     const client = MatrixClientPeg.get();
     if (client.isGuest()) {
-        setTimeout(()=>{
-            Modal.createDialog(NeedToRegisterDialog, {
-                title: _t('Please Register'),
-                description: _t('Guest users can\'t create new rooms. Please register to create room and start a chat.')
-            });
-        }, 0);
+        dis.dispatch({action: 'view_set_mxid'});
         return q(null);
     }
 
@@ -62,6 +56,11 @@ function createRoom(opts) {
     }
     if (opts.dmUserId && createOpts.is_direct === undefined) {
         createOpts.is_direct = true;
+    }
+
+    // By default, view the room after creating it
+    if (opts.andView === undefined) {
+        opts.andView = true;
     }
 
     // Allow guests by default since the room is private and they'd
@@ -97,10 +96,12 @@ function createRoom(opts) {
         // room has been created, so we race here with the client knowing that
         // the room exists, causing things like
         // https://github.com/vector-im/vector-web/issues/1813
-        dis.dispatch({
-            action: 'view_room',
-            room_id: roomId
-        });
+        if (opts.andView) {
+            dis.dispatch({
+                action: 'view_room',
+                room_id: roomId,
+            });
+        }
         return roomId;
     }, function(err) {
         console.error("Failed to create room " + roomId + " " + err);